+++ src/server/scripts/Pandaria/GateOfTheSettingSun/boss_striker_gadok.cpp
