+++ src/server/scripts/Pandaria/GateOfTheSettingSun/instance_gate_of_the_setting_sun.cpp
